6-size.c

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long long int d;
 	float f;
 
-printf("Size of a char:%d byte(s)\n", (unsigned long)sizeof(a));
-printf("Size of a int:%d byte(s)\n", (unsigned long)sizeof(b));
-printf("Size of a long int:%d bytes(s)\n", (unsigned long)sizeof(c));
-printf("Size of a long long int:%d bytes(s)\n", (unsigned long)sizeof(d));
-printf("Size of a float:%d bytes(s)\n", (unsigned long)sizeof(f));
+printf("Size of a char:%d byte(s)\n", sizeof(a));
+printf("Size of a int:%d byte(s)\n", sizeof(b));
+printf("Size of a long int:%d bytes(s)\n", sizeof(c));
+printf("Size of a long long int:%d bytes(s)\n", sizeof(d));
+printf("Size of a float:%d bytes(s)\n", sizeof(f));
 return (0);
 }
